float instead of bd

+++ lib/private_api.rb
@@ -40,8 +40,8 @@ module PrivateApi
     params = {
       pair: pair,
       type: order_type,
-      volume: volume.to_d.round(6).to_s,
-      price: price.to_d.round(6).to_s
+      volume: volume.to_f.round(6).to_s,
+      price: price.to_f.round(6).to_s
     }
     opt.merge!({params: params, method: :post})
     path = '/api/1/postorder'
